reader.py: refactoring, add readInFile()
refactored the functions operating on a ZipFile object to operate on the
path of the extracted zip file instead.
readInFile() combines everything releated to reading in a bcf file. It
runs the validation against the schema files as well as the building of
the objects (like a Viewpoint object for example). Till now
buildViewpoint() and buildMarkup() are still placeholder methods.