Add src/bcf/test_data with valid topics
This folder is different form the one for the unit tests. It is intended
to be used for testing of the functions before unit tests are written
for them.
Add file generate_zip.fish which takes a number of topic directory names
(which are UUIDs) inside an extracted BCF file, and compresses them in
conjunction with project.bcfp and bcf.version to a BCF compliant zip
file.