Bug fixes due to testing
__str__ function was implemented for some classes, to be better able to
debug/trace failing test cases.
Additionally in reader comments now get the right viewpoint reference
assigned in a second step inside `buildMarkup()`.
Bug fixes in reader.py
reader.getViewpointRefByGuid() was added for the second step in
buildMarkup() in which the comments get assigned the right reference to
the viewpoints. Before this step is executed only a temporary
ViewpointReference object exists that holds just the Viewpoint Guid.