markup.py: comment inherits Identifiable
Comment also accepts guid in the init function as required argument.
reader.py and reader-tests.py updated accordingly
Bugfix in util.py: changed fileURL to fileUrl
writer.py:  finished. Reads out the hierarchy of an element and checks
whether it is part of an other element that occurs in a list. If that is
the case then the Id of that list element is returned.