project.py: Add SimpleElement, SimpleList, Attribute
These classes are used to replace the elementary types of member
variables representing a simple element, attribute or a list of simple
elements. Every simple element also inherits form XMLName and Hierarchy.
Elements in viewpoint.py do not contain SimpleElements, Attributes or
SimpleLists since a viewpoint is assumed to be added just once and only
once the XMLString has to be generated, which will be done in Viewpoint
itself.
Every class containing SimpleElement, Attribute or a SimpleList also
contains a property for the affected member variable to enable direct
access to the encapsulated value.