writer.py: viewpoints can now be added
if a new viewpoint reference is added addElement() automatically checks
whether the referenced viewpoint is also new. If so not only markup.bcf
is edited but also a new viewpointX.bcfv file is created with the
contents of the viewpoint object serialized.

Additionally the outputs of addElement() are now nicely formatted.
Therefore xmlPrettify() was added.