writer.py: compileChanges() function added
In addition to this function the function getStateList() as added to the interface
State. This function returns a list of tuples where each tuple contains
the state of an object and the object itself, it it is not in its
original state. This function is utilized by writer.compileChanges() to
get a list of objects whose state requires an action.