Add two test cases. Rewrite getInsertionIndex()
The testcases that were added are:
  - add a file attribute: this test case is currently failing
  - add a file node in header: this test case, as well as the others
  runs through.

writer.getInsertionIndex(): the algorithm is rewritten to always return
the greatest most index at which the new element could be inserted while
complying with the schema file. Thus if multiple nodes of the same kind
element becomes, the node (resulting from element) would be inserted
last in row.