BugFix + Refactor assignments of containingObject
Previously the containingObject was assigned in reader. Now it is moved
into the constructors of the classes itself, which makes more sense imo.

BugFix: if a class inherited from Identifiable then the equality check
of id was not done properly. Therefore identifiable implements the
function idEqual, which checks two for equality.