Rename and extend getContainingElementForAttribute
writer.getContainingElementForAttribute is now
writer.getEtElementFromFile. It thus shall symbolise that it is a more
general form than the previous version. The function can search for
elements in general in a given xmlroot. The old version got extended by
a third matching strategy, in which elements are matched against each
other based on their texts.
It also now accepts a third parameter `ignoreNames`. Every element
attribute, whose name is equal to an entry in `ignoreNames` is ignored
in the matching phase. For example the following two elements would be
equal with `ignoreNames = [ "hello" ]`:

  <File hello="world" goodbye="world">
    <hello/>
    <bye/>
  </File>
  <File goodbye="world"/>
    <bye/>
  </File>

Note that `ignoreNames` is applied to both attributes and child elements
of File.

reader.py: changed the type of IfcSpatialStructureElement from UUID to
string, as it is not intended to be an unique identifier for File.

writer_tests.py: add new group of testcases `GetEtElementFromFileTests`.
The contained tests check the correct function of
writer.getEtElementFromFile().
Move `setupBCFFile` and `compareFiles` from `writer.AddElementTests`
into the global scope.
The added test cases in particular are:
- test_findComment: A comment is searched for by its children in a file
  that contains a similar comment with two children missing.
- test_findComment2: A comment is searched for by its children in a file
  that contains a similar comment in which only the text of one child
  differs by one character from the other one.
- test_findFileByAttribute: A file node is searched for that does not
  contain text or children, only attributes.
- test_findLabelByText: A label node is searched for that only differs
  in its text from the other present label nodes.