Refactor interfaces.Identifiable
The id member variable is now always of type UUID. This change is more
intutive with the standard, as only elements with a Guid property can be
identified by it. All other elements that have an Id-like
property/attribute use it to reference another uniquely identified
object. Hence all other Id-like elements may occur more than once in a
bcf file, which is against the intended purpose of the Identifiable interface.