util.py: add updateSchemas(), copySchemas()
updateSchemas downloads the schema files from github and stores them in
PLUGIN_ROOT/src/schemas/. This folder is created if it does not exists.

copySchemas first checks whether the schemas already exist, if not
calls updateSchemas(). After that has happened the files are copied to
the supplied destination directory and the file paths are returned as
tuple