writer.py: add modifyElement, processProjectUpdate
modifyElement() does the job of, well, modifying an element. Thereby an
element can only be of type `Attribute` or `SimpleElement`, all other
types have members that are of one of the modifiable types or also of
complex type. In grammar this would be

element =  Attribute
         | SimpleElement
         | Member, Member
Member = element

processProjectUpdate() walks through the also added list
`projectUpdates` and processes each with the respective handler
function. An element `projectUpdates` is a tuple consisting of a deep
copy of the updated project object, the object that was changed and the
old value of the object if object is of type `SimpleElement` or
`Attribute`, otherwise `None`. Depending on the state of the object
handleAddElement(), handleDeleteElement() or handleModifyElement() will
be called. These functions handle all errors addElement(),
deleteElement() or modifyElement() may raise.

reader.py: now checks whether the download of the schema files was
successful, and if not notifies the user through an exception. Will be
handled in the UI in the future.

writer_tests.py: three testcases were added for testing modifyElement()
The testcases are:
  - test_modifyFileName: modifies the text of a simple xml element
  - test_modifyIfcProject: modifies the value of a xml attribute
  - test_modifyComment: modifies the value of the comment element of
    a comment