identifiable.py: Add new Identifiable class
identifiable.Identifiable has been renamed to
identifiable.XMLIdentifiable as it only holds ids assigned to elements
in the xml file (i.e.: the values of the guid attributes).

A new identifiable.Identifiable class was added. It shall provide every
inheriting class with a unique id that gets generated at object creation
and shall not be changed anymore during the lifetime of the object. This
enables efficient search algorithms that comb through the complete
datamodel. This id also shall not be changed after a deep copy of an
object.