Implement searchObject(). Add search_tests.py
searchObject() is integrated in every Identifiable class in the data
model. Its purpose is to check if itself has the same id as the supplied
object (if this is Identifiable). If not then the function is supposed
to check its members for the supplied object's id.

To check the functionality of searchObject() unit tests were also added
that twice search for an element and once for an attribute