interface_tests.py: add testcases for frontentInterface.deleteObject()
the function deleteObject() is finished, at least I have a great deal of confidence that no bugs in this particular implementation persist. 
The test-suit however shows some strange behavior. If all tests are run at once then some test case fail. But when the failing test cases are run independently they succeed. This will be fixed in the next commit on this branch.