topic.py: Rename refs -> docRefs; fInterface.py: use correct addUpdate()
the member variable 'refs' was renamed to docRefs to make it clearer that it holds a list of document references. 

frontentInterface.py: prior to this commit it used a writer.addUpdate() to add a new update to projectUpdates. this was changed to writer.addProjectUpdate() since it is safer to use.