writer.py: add createBcfFile(); util.py: add cd class
createBcfFile creates a new BCF file complying with the standard. It packs the contents of the supplied directory into an archive. Thereby is every directory itself also stored as single entry. Beneath earch directory the file contained in this directory are stored. 

util.py: add a context manager for os.chdir to savely change directory. currently used only for creating the bcf file.