pI: add modifyComment(), writer: refactor getEtElementFromFile()
writer.getEtElementFromFile() used a pretty simple XMLPath expression to get a list of possible candidates for matching against the searched element. That was problematic in the case of xml elementst that occured twice in the hierarchy, like ModifiedAuthor or ModifiedDate. To solve a new function is introduced writer.getXMLNodeCandidates() that constructs a XMLPath expression matching the complete hierarchy of the element. Currently it only supports elements inside markup.bcf, if they are not attributes.