CommentView finished in v0.1
The list of comments is done with a view (to display the data), a model (that provides access to the data) and a delegate (that handles the editing of data). 

The view is contained in ./bcfplugin/gui/comment-list/mainwindow.py
The model is contained in ./bcfplugin/gui/comment-list/commentmodel.py
The delegate is contained in ./bcfplugin/gui/comment-list/commentdelegate.py

Currently a double click triggers the creation of an editor. In this editor the comment and the author can be changed. In later versions probably a separate widget will be created.