BugFix: draw at wrong position, disappear on select
The drawing of comments starts at `topY`, the top y coordinate, and then continues downwards. This variable was calculated by the class itself, instead of using `option.rect` to get the top Y coordinate. Much more reliable. 
This also fixed the bug of a disappearing comment when selected, which in the end was caused by a wrong top Y coordinate.